+++ src/factory/reservation/event.ts
@@ -2,8 +2,7 @@ import {
     ICOAInfo,
     ILocation as IEventLocation,
     IName as IEventName,
-    ISuperEvent,
-    IWorkPerformed
+    ISuperEvent
 } from '../event/screeningEvent';
 import { EventStatusType } from '../eventStatusType';
 import { EventType } from '../eventType';
@@ -16,7 +15,6 @@ import * as ReservationFactory from '../reservation';
 import { ReservationType } from '../reservationType';
 
 // IReservationForを最適化
-// export type IReservationFor = IScreeningEvent;
 export interface IReservationFor {
     endDate: Date;
     eventStatus: EventStatusType;
@@ -36,7 +34,8 @@ export interface IReservationFor {
     startDate: Date;
     superEvent: ISuperEvent;
     typeOf: EventType.ScreeningEvent;
-    workPerformed?: IWorkPerformed;
+    // ↓廃止(2022-04-18~)
+    // workPerformed?: IWorkPerformed;
     doorTime?: Date;
 }